Update Dead.java
Addition comments:
Currently works to utilize the Tiny BoxBitA10 font, however in its current state returns an runtime error which doesn't seem to interrupt the program running.
Also requires the font file to be placed in the same location as the code. The file path of the font must be the same as the file path of the program.